requirements.txt
pandas
duckdb
openpyxl
langchain
langchain-openai
python-dotenv
